Fix concerns, about security, also instruct user in README.md
