+++ gradle/wrapper/gradle-wrapper.properties
@@ -5,4 +5,4 @@ zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
 org.gradle.jvmargs=-Xmx2048m -XX\:MaxPermSize\=2048m -Xms512m
 org.gradle.daemon=true
-distributionUrl=https\://services.gradle.org/distributions/gradle-2.4-all.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.8-all.zip
